+++ docs/src/utils/mockGraphql.ts
@@ -1,12 +1,17 @@
 import fetchMock from 'fetch-mock';
+import resource from '../../../src/spec/mock/elegant-cms/resource';
 
-let realFetch: (input: RequestInfo, init?: RequestInit) => Promise<Response>;
-if (typeof window !== 'undefined') {
-  // @ts-ignore
-  realFetch = fetch;
-}
-
-const waitForDuration = (time: number) => new Promise(resolve => setTimeout(resolve, time));
+const resourcesMock = {
+  data: {
+    resources: {
+      pageInfo: {
+        hasNextPage: false,
+        endCursor: 'spookyboo',
+      },
+      edges: [{ node: resource }, { node: resource }, { node: resource }, { node: resource }],
+    },
+  },
+};
 
 const credentialResourceMock = {
   data: {
@@ -18,17 +23,30 @@ const credentialResourceMock = {
   },
 };
 
+let realFetch: (input: RequestInfo, init?: RequestInit) => Promise<Response>;
+if (typeof window !== 'undefined') {
+  // @ts-ignore
+  realFetch = fetch;
+}
+
+const waitForDuration = (time: number) => new Promise(resolve => setTimeout(resolve, time));
+
 // TODO: Modify the gatsby plugin so we can fetch this data at build time
 export const mockGraphQl = () => {
   fetchMock.mock('express:/graphql', async (url, opts) => {
     await waitForDuration(300);
 
-    // Mocks manifold-credentials component.
-    if (opts && opts.body && String(opts.body).includes('query RESOURCE_CREDENTIALS')) {
+    const { headers, body } = opts;
+
+    const bodyString = body as string;
+    // TODO(dom): This ok right now but as we scale out our queries and such we should
+    // use a proper gql mocking tool that supports better field mocking ect.
+    if (bodyString.includes('query RESOURCES')) {
+      return resourcesMock;
+    }
+    if (bodyString.includes('query RESOURCE_CREDENTIALS')) {
       return JSON.stringify(credentialResourceMock);
     }
-
-    const { headers } = opts;
     // @ts-ignore
     delete headers.authorization;
     const result = await realFetch(url, {